Introduce communication for projects and other resources
This orients itself on the ResourcesService that was introduced int he offer manager.
However, it is improved in the regards that the information on what change was made is now also communicated.
Furthermore the event emmitter is not necessary and you only have one service/channel on which you publish changes.